requirements.txt
python-dotenv>=1.0.1
langchain>=0.2.0
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-postgres>=0.0.14
langchain-experimental>=0.1.0
langchain-graph-retriever>=0.1.0
neo4j>=5.20.0
tiktoken>=0.6.0
psycopg[binary]>=3.1.19
streamlit>=1.28.0
pyvis>=0.3.2
pymupdf>=1.24.0
streamlit-agraph>=0.0.45
networkx>=3.0
sudachipy>=0.6.8
sudachidict_core>=20240716